# api/openapi.yaml
openapi: 3.0.3
info:
  title: TalentSphere API
  description: |
    Comprehensive Learning Management System API for course management, 
    user progress, challenges, certificates, and AI-powered learning.
    
    ## Authentication
    All endpoints (except health checks) require JWT Bearer token authentication.
    
    ## Rate Limiting
    100 requests per minute per authenticated user.
    
    ## Versioning
    Current version: v1.0
    
    ## Support
    API Documentation: https://api.talentsphere.com/docs
    Support: api-support@talentsphere.com
  version: 1.0.0
  contact:
    name: TalentSphere API Team
    email: api-support@talentsphere.com
    url: https://talentsphere.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.talentsphere.com/v1
    description: Production server
  - url: https://staging-api.talentsphere.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        field:
          type: string
          description: Field that caused the error (if applicable)
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        details:
          type: object
          description: Additional error details

    # Auth Schemas
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password (min 8 characters)
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, ADMIN]
          default: STUDENT
          description: User role
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          description: User password
    AuthResponse:
      type: object
      required: [userId, email, role, accessToken, expiresIn]
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: User email
        role:
          type: string
          enum: [STUDENT, INSTRUCTOR, ADMIN]
          description: User role
        accessToken:
          type: string
          description: JWT access token
        expiresIn:
          type: integer
          description: Token expiration time in seconds

    # Course Schemas
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...