# backends/backend-springboot/src/main/java/com/talentsphere/backend/controller/AuthController.java
package com.talentsphere.backend.controller;

import com.talentsphere.backend.dto.AuthResponse;
import com.talentsphere.backend.dto.LoginRequest;
import com.talentsphere.backend.dto.RegisterRequest;
import com.talentsphere.backend.service.AuthService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/auth")
public class AuthController {

    @Autowired
    private AuthService authService;

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request) {
        return new ResponseEntity<>(authService.register(request), HttpStatus.CREATED);
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}